requirements.txt
Django==2.2.10
psycopg2-binary==2.8
djangorestframework==3.11.0
ipython
requests